requirements.txt
streamlit==1.28.1
pandas==2.1.3
mysql-connector-python==8.2.0
openpyxl==3.1.2
xlrd==2.0.1